[BlockBroadcastRefactor] Refactoring and renaming
